feat: make a number of QoL improvements
- make the package into a module
- add nix flakes to ease development
- include poetry
- edit the packet structure
- use model based approach
