+++ .gitignore
@@ -1,5 +1,4 @@
 *.txt
 *.json
-!requirements.txt
-venv
-
+.direnv
+.envrc
