+++ README.md
@@ -1,14 +1,32 @@
 # nrich.py
 
-Inspired on the [nrich](https://gitlab.com/shodan-public/nrich) tool developed by Shodan, but alas that tool was in Rust.
-
-Thus I have made a Python3 port of that tool, to make it more easy to include in other projects.
+Python equivalent of the rust [nrich](https://gitlab.com/shodan-public/nrich) program, developed by Shodan. Simple async Python program to interact with the Shodan InternetDB API.
 
 In addition to the official tool, you can specify subnets, and this tool just unpacks them.
 
-# parse_internetdb.py
+## Usage
 
-Example script to parse the shodan internet.db output, using simple for looping and if-ing.
+### nrich.py
+
+`nrich --help`:
+
+```
+usage: nrich.py [-h] [-input-file INPUT_FILE] [-output-file OUTPUT_FILE] [-output-type {json,str}] [-skip_missing]
+
+options:
+  -h, --help            show this help message and exit
+  -input-file INPUT_FILE, -i INPUT_FILE
+                        File to parse IPs from, if not given parse from stdin
+  -output-file OUTPUT_FILE, -o OUTPUT_FILE
+                        file to write IPs to
+  -output-type {json,str}, -t {json,str}
+                        Format to output in
+  -skip_missing, -s     skip IPs with no information on them.
+```
+
+### parse_internetdb.py
+
+Example script to parse the shodan internet.db output
 
 # Installation
 
@@ -18,6 +36,12 @@ Make a virtual environment and install dependencies, like so:
 python3 -m venv venv
 source ./venv/bin/activate
 pip install -r requirements.txt
+```
+
+## Nix users
+
+```
+nix develop
 ```
 
 # License
