+++ result
@@ -1,0 +1,1 @@
+/nix/store/h27qxk9ks8137klpgbhaf1chpvxjz779-python3.11-nrich-0.1.0
