+++ src/nrich/__init__.py
@@ -1,0 +1,3 @@
+from .models import ShodanResult
+
+__all__ = ["ShodanResult"]
